Bs4RedisAnsible/src/requirements.txt
async-timeout==4.0.3
autopep8==2.0.4
beautifulsoup4==4.12.2
bs4==0.0.1
pycodestyle==2.11.1
PyYAML==6.0.1
redis==5.0.1
soupsieve==2.5
tomli==2.0.1
